fix(sync): apply the baseline leniency to force-pull conflict detection
`sync pull --force` aborts with SYNC_CONFLICT when a config is both
locally modified and changed on the remote. It compared the stored
baseline strictly, so an entry written before the script-shape change
(issue #686) read as "remote changed" and a force-pull aborted on a
config nobody had touched remotely -- the third place a stored hash
meets a fresh API hash, after the diff base and the pull idempotency
check.

`_is_conflict` and `_detect_force_pull_conflicts` move to
`_sync_baseline` as free functions, joining the other two comparison
sites, and the config-level check now runs the stored hash through
`effective_stored_hash`. Rows stay strict: their hash producer never
changed. sync_service.py shrinks by ~78 code lines in the process.
